TransientSpec 1D class
