+++ lumispy/hyperspy_extension.yaml
@@ -150,8 +150,18 @@ signals:
     lazy: True
     module: lumispy.signals.luminescence_transient
 
+  TransientSpectrumCasting:
+    signal_type: TransientSpec
+    signal_type_aliases:
+      - TRLumiSpec
+      - TR luminescence spectrum
+      - time-resolved luminescence spectrum
+    signal_dimension: 1
+    dtype: real
+    lazy: False
+    module: lumispy.signals.luminescence_transientspec
   LumiTransientSpectrum:
-    signal_type: Luminescence
+    signal_type: TransientSpec
     signal_type_aliases:
       - TRLumiSpec
       - TR luminescence spectrum
@@ -161,9 +171,8 @@ signals:
     lazy: False
     module: lumispy.signals.luminescence_transientspec
   LazyLumiTransientSpectrum:
-    signal_type: Luminescence
+    signal_type: TransientSpec
     signal_type_aliases:
-      - TransientSpec
       - TRLumiSpec
       - TR luminescence spectrum
       - time-resolved luminescence spectrum
